requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn>=0.34.0,<1.0.0
requests>=2.32.0,<3.0.0
openai>=1.68.0,<2.0.0
google-genai>=0.3.0,<1.0.0
supabase>=2.15.0,<3.0.0
pydantic-settings>=2.8.0,<3.0.0
pytest>=8.3.0,<10.0.0
httpx>=0.28.0,<1.0.0
